backend/requirements.txt
fastapi[standard]==0.119.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
python-dotenv==1.0.1
gunicorn==21.2.0
python-slugify==8.0.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
bcrypt==4.0.1
httpx==0.27.0
